add bar image

+++ StayGo.Domain/Models/Bar.cs
@@ -8,7 +8,7 @@ namespace StayGo.Domain.Models
         public string Address { get; set; }
         public string Description { get; set; }
         public string Type { get; set; }
-        public byte Rating { get; set; }
+        public int Rating { get; set; }
         public double AvgPayCheck { get; set; }
     }
 }
