+++ StayGo.Server/web/Models/BarDto.cs
@@ -10,7 +10,7 @@ namespace staygo_server.Models
         public string Address { get; set; }
         public string Description { get; set; }
         public string Type { get; set; }
-        public byte Rating { get; set; }
+        public int Rating { get; set; }
         public double AvgPayCheck { get; set; }
     }
 }
